[CAMEL-11807] Upgrade camel-core to junit5

+++ core/camel-core/src/test/java/org/apache/camel/converter/NIOConverterTest.java
@@ -22,7 +22,10 @@ import java.nio.ByteBuffer;
 
 import org.apache.camel.ContextTestSupport;
 import org.apache.camel.Exchange;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class NIOConverterTest extends ContextTestSupport {
 
@@ -124,7 +127,7 @@ public class NIOConverterTest extends ContextTestSupport {
     public void testToByteBufferFloat() {
         ByteBuffer bb = NIOConverter.toByteBuffer(Float.valueOf("2"));
         assertNotNull(bb);
-        assertEquals(2.0f, bb.getFloat(), 1e-5f);
+        assertEquals(2.0f, bb.getFloat(), (double) 1e-5f);
     }
 
     @Test
